+++ README.md
@@ -11,4 +11,4 @@
 ## 3.JSON View 구현하기
 - [x] JspView 클래스 구현
 - [x] JsonView 클래스 구현
-- [ ] Legacy MVC 제거
+- [x] Legacy MVC 제거
