+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/HandlerExecutionAdapter.java
@@ -1,8 +1,7 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.ModelAndView;
 import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerExecution;
 
 public class HandlerExecutionAdapter implements HandlerAdapter {
