+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/HandlerMappings.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet;
 
 import jakarta.servlet.http.HttpServletRequest;
 import webmvc.org.springframework.web.servlet.mvc.tobe.AnnotationHandlerMapping;
